+++ clients/admin-ui/src/features/common/custom-fields/CreateCustomLists.tsx
@@ -5,10 +5,10 @@ import * as Yup from "yup";
 
 import { getErrorMessage } from "~/features/common/helpers";
 import { useAlert } from "~/features/common/hooks";
+import { TrashCanSolidIcon } from "~/features/common/Icon/TrashCanSolidIcon";
 import { useUpsertAllowListMutation } from "~/features/plus/plus.slice";
 import { AllowListUpdate } from "~/types/api";
 
-import { TrashCanSolidIcon } from "../Icon/TrashCanSolidIcon";
 import CustomInput from "./form/CustomInput";
 import { AddIcon } from "./icons/AddIcon";
 import { Layout } from "./Layout";