+++ clients/admin-ui/src/features/user-management/UserForm.tsx
@@ -13,7 +13,6 @@ import {
 import { SerializedError } from "@reduxjs/toolkit";
 import { FetchBaseQueryError } from "@reduxjs/toolkit/dist/query";
 import { useAPIHelper } from "common/hooks";
-import { TrashCanSolidIcon } from "common/Icon/TrashCanSolidIcon";
 import { Form, Formik } from "formik";
 import NextLink from "next/link";
 import React from "react";
@@ -24,6 +23,7 @@ import { useAppDispatch, useAppSelector } from "~/app/hooks";
 import { USER_MANAGEMENT_ROUTE } from "~/constants";
 import { CustomTextInput } from "~/features/common/form/inputs";
 import { passwordValidation } from "~/features/common/form/validation";
+import { TrashCanSolidIcon } from "~/features/common/Icon/TrashCanSolidIcon";
 import { successToastParams } from "~/features/common/toast";
 
 import PasswordManagement from "./PasswordManagement";